third_party/blink/web_tests/platform/generic/fast/events/event-input-contentEditable-expected.txt

This tests that the input events are dispatched when contentEditable nodes are edited

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.target.id is 'target0'
PASS event.target.innerHTML is 'Text'
PASS event.target.id is 'target2'
PASS event.target.innerHTML is 'This text should not be changed.'
PASS event.target.id is 'target3'
PASS event.target.innerHTML is ''
PASS event.target.id is 'target4'
PASS event.target.innerHTML is '<a href="http://www.example.com/">This text should be a link.</a>'
PASS event.target.id is 'target6parent'
PASS event.target.innerHTML is '<a href="http://www.example.com/"><span id="target6start">Start,</span><span id="target6middle">Middle,</span><span id="target6end">End.</span></a>'
PASS event.target.id is 'target7'
PASS event.target.innerHTML is 'X'
PASS event.target.id is 'target8'
PASS event.target.innerHTML is ''
PASS event.target.id is 'target9parent'
PASS event.target.innerHTML is '<div id="target9child" contenteditable="">Replacing</div>'
PASS event.target.id is 't10gch'
PASS event.target.innerHTML is 'Replacing'
PASS expectedInputEventCount is actualInputEventCount
PASS successfullyParsed is true

TEST COMPLETE
